doc: test examples
The examples of the tutorials have to be tested so they are moved into file so
the test and the documentaion can include them.

*docs/examples*

The example snippets are moved into separated directory under `docs/examples`.
The example directories has the following structure:
 * `main.py`: the demo scriptof an argparse_action featues
 * exmaple calls
   ```
   $ python3 main.py ...
   <answer>
   ```

*tests/test_examples.py*

`test_examples` loads the test modules (docs/exmaples/*/main.py) and checks the
exmaple calls. `test_examples` could executed the example calls with
`subprocess.run`, but it would require to invoke a virtualenv where the
`argparse_action` is already installed. With the module load the tests can be
executed without the installation of `argparse_action`

`test_examples` evaluates the example calls with `subTest` so they will be
handled as individual test steps by the unittest test framework.