argparse_action: pass argparse options via keyword arguments #3
argparse options can be passed to the arguments of the decorated function
via keyword arguments if `Action.add` ot `add_action`. The passed options
extend the options which are parsed from the function signature, so whether
the CLI argument is positional or optional will be decided from the default
value of the function argument.