test/test_examples: use suprocess to check examples #3
Module load was not reliable about the name of the running program.
The test_examples wants to ensure that all the examples in the docs
are correct. That can be evaulated with a simple `subprocess.run` call.

The test_examples runs the examples with the .venv/bin/python interpreter
where the argparse_command is installed.