argparse_action: handles underscore in parameter names #6
argparse converts '-' to '_' in the name of the optional arguments but the name
of the positional arguments are untouched. Therefore `_wrap_action` fetch the
posotional arguments (args, vargs) with the cli formatted name (s/_/-/g).

The parameters with default values are handled as positional function parameter
but they are converted to optional cli argument. Therefore `_wrap_action`
converts the parameter name to cli style only if the parameter did not have
default value.