argparse_action/logging: add_log_arguments, init_logging #7
`add_log_arguments` creates options for the standard python logging module. Log
destination can be syslog, file or console and the logging can be disabled with
`--log-none` option. The default log destination is console.

The log messages will be sent with "user" facility into syslog.

The default log level is "info" which can be overriden with `--log-level`
option.

The log cli arguments are registered into the main argparse parser so they have
to be specified before the command selection.