argparse_action.action: argparse append action for sequence default values
A CLI option can be defined multiple times. If the default value of the
parameter is a sequence, then argparse_action will invoke an argparse
"append" action so the CLI options will be collected into a list.

The type annotation of the sequence default value should be wrapped into
`collection.abc.Sequence[<type>]`.